moved atomic mass dict to a separate submodule so it is only instantiated once and adjusted tests to account for this
